DRF-JWT-Djoser/drf-ultimate/drf-ultimate/requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
cloudinary==1.44.1
Django==5.2.7
django-cloudinary-storage==0.3.0
djangorestframework==3.16.1
idna==3.11
pillow==12.0.0
requests==2.32.5
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
